# swagger.yaml
openapi: 3.0.0
info:
  title: Inventory API
  version: 1.0.0
servers:
  - url: /
paths:
  /inventory:
    get:
      summary: Get all items
      responses:
        '200':
          description: List of items
  /register:
    post:
      summary: Register new item
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                inventory_name:
                  type: string
                description:
                  type: string
                photo:
                  type: string
                  format: binary
      responses:
        '201':
          description: Created
  /search:
    post:
      summary: Search item
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: string
                has_photo:
                  type: string
      responses:
        '200':
          description: Found
  /inventory/{id}:
    get:
      summary: Get item details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Item details
        '404':
          description: Not Found
    put:
      summary: Update item details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Updated
    delete:
      summary: Delete item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted